feat: convert forgotten pages to typescript
