+++ showcase/app/routes/application.ts
@@ -5,9 +5,10 @@
 
 import Route from '@ember/routing/route';
 import { service } from '@ember/service';
+import type IntlService from 'ember-intl/services/intl';
 
 export default class ApplicationRoute extends Route {
-  @service intl;
+  @service declare intl: IntlService;
 
   beforeModel() {
     this.intl.setLocale(['en-us']);
