--- showcase/app/routes/page-foundations.js
@@ -1,8 +1,0 @@
-/**
- * Copyright (c) HashiCorp, Inc.
- * SPDX-License-Identifier: MPL-2.0
- */
-
-import Route from '@ember/routing/route';
-
-export default class PageFoundationsRoute extends Route {}
